Allow multiple instruments playing on one track
Instead of having a single sampler play multiple sounds, allow a sampler
to process audio for multiple tracks. This makes it possible to combine
different kinds of instruments on a single track.

Also defines a plugin trait that simplifies the engine and should make
it easier to support external plugins later.